+++ fw/lyfi/main/src/solar.h
@@ -23,20 +23,21 @@ enum solar_instants_enum {
  * @brief Structure representing a key point in the solar day.
  */
 struct solar_instant {
-    double time; // Time in hours (e.g., 6.5 represents 6:30)
-    double brightness; // Brightness value (0 to 1.0)
+    float time; // Time in hours (e.g., 6.5 represents 6:30)
+    float brightness; // Brightness value (0 to 1.0)
 };
 
 /**
  * @brief Calculates the timezone offset in hours for a given date.
  * @return The timezone offset in hours.
  */
-double solar_calculate_timezone_offset(const struct tm* tm_local);
+float solar_calculate_timezone_offset(const struct tm* tm_local);
 
 /**
  * @brief Calculate sunrise and sunset times
  * @param latitude Location latitude (-90 to 90)
  * @param longitude Location longitude (-180 to 180)
+ * @param timezone_offset Timezone offset in hours
  * @param tm_local Local time
  * @param[out] sunrise Calculated sunrise time (hours, 0-24)
  * @param[out] sunset Calculated sunset time (hours, 0-24)
@@ -47,18 +48,17 @@ double solar_calculate_timezone_offset(const struct tm* tm_local);
  *  -ENODATA: Polar night/midnight sun condition
  *  -EFAULT: Time calculation error
  */
-int solar_calculate_sunrise_sunset(double latitude, double longitude, double timezone_offset, const struct tm* tm_local,
-                                   double* sunrise, double* sunset);
+int solar_calculate_sunrise_sunset(float latitude, float longitude, float timezone_offset, const struct tm* tm_local,
+                                   float* sunrise, float* sunset);
 
 /**
  * @brief Generates key points for solar brightness throughout the day.
  * @param sunrise The sunrise time in hours.
  * @param sunset The sunset time in hours.
  * @param instants Array to store the generated instants.
- * @param max_points Maximum number of key points to generate.
- * @return The number of key points generated.
+ * @return 0 on success
  */
-int solar_generate_instants(double sunrise, double sunset, struct solar_instant* instants);
+int solar_generate_instants(float sunrise, float sunset, struct solar_instant* instants);
 
 #ifdef __cplusplus
 }
